Split Source constants into a new file
This will make it slightly easier to generate automatic updates to the
Source list when a new prebuilt is released, since the whole
`source_constants.rs` file can be generated, rather than modifying a
section of `lib.rs` in-place.
